Moved config manager to separate class
